🎨 Reorg specs
